+++ lib/src/commands/commands.dart
@@ -1,2 +1,4 @@
+export 'analyze_command.dart';
+export 'fix_command.dart';
 export 'sample_command.dart';
 export 'update_command.dart';